Moved formatRoles into the codebase & GetRolesForToken Strictly Checks for JWTRoles return value
